move data utils to their own module; add function to check for valid times and shift
